+++ src/app/angular-material.module.ts
@@ -1,0 +1,71 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import {
+   MatButtonModule,
+   MatCardModule,
+   MatToolbarModule,
+   MatIconModule,
+   MatBadgeModule,
+   MatSidenavModule,
+   MatListModule,
+   MatGridListModule,
+   MatFormFieldModule,
+   MatInputModule,
+   MatSelectModule,
+   MatRadioModule,
+   MatDatepickerModule,
+   MatNativeDateModule,
+   MatChipsModule,
+   MatTooltipModule,
+   MatTableModule,
+   MatPaginatorModule
+} from '@angular/material';
+
+@NgModule({
+   imports: [
+      CommonModule,
+      MatButtonModule,
+      MatCardModule,
+      MatToolbarModule,
+      MatIconModule,
+      MatSidenavModule,
+      MatBadgeModule,
+      MatListModule,
+      MatGridListModule,
+      MatFormFieldModule,
+      MatInputModule,
+      MatSelectModule,
+      MatRadioModule,
+      MatDatepickerModule,
+      MatNativeDateModule,
+      MatChipsModule,
+      MatTooltipModule,
+      MatTableModule,
+      MatPaginatorModule
+   ],
+   exports: [
+    MatCardModule,
+      MatButtonModule,
+      MatToolbarModule,
+      MatIconModule,
+      MatSidenavModule,
+      MatBadgeModule,
+      MatListModule,
+      MatGridListModule,
+      MatInputModule,
+      MatFormFieldModule,
+      MatSelectModule,
+      MatRadioModule,
+      MatDatepickerModule,
+      MatChipsModule,
+      MatTooltipModule,
+      MatTableModule,
+      MatPaginatorModule
+   ],
+   providers: [
+      MatDatepickerModule,
+   ]
+})
+
+export class AngularMaterialModule { }