+++ CHANGELOG.md
@@ -6,4 +6,5 @@
 
 - Lavoro sul branch feature/auth per la creazione della mappa di login
 - Aggiunto Material (ng add @angular/material)
-- Aggiunto Bootstrap (npm install bootstrap)
+- Aggiunto Flex Layout (npm i -s @angular/flex-layout@8 @angular/cdk)
+- Aggiunto il modulo angular-material.module.ts